Task#3
Write a program in which you have to remove the ID section from the Data (IRIS) and then find the correlation of all the data and print them. At last, create a Heatmap to show their relation with green, blue color palette and make sure the annotations are turned on.